requirements.txt
# Core ML
numpy>=1.24
pandas>=2.0
scipy>=1.11
scikit-learn>=1.3
xgboost>=2.0
lightgbm>=4.0

# Tracking
mlflow>=2.8
optuna>=3.4

# API
fastapi>=0.104
uvicorn>=0.24
pydantic>=2.4

# Config
pyyaml>=6.0
python-dotenv>=1.0

# Visualization
matplotlib>=3.8
seaborn>=0.13